+++ src/types/utility-types/literal.ts
@@ -33,20 +33,36 @@ export class Literal<T> extends Type<T, T> {
  * The literal type will return a type that will match only the exact given type.
  * The given value must be a primitive, in order to be serialized to a snapshot correctly.
  * You can use literal to match exact strings for example the exact male or female string.
- * 
+ *
  * @example
  * const Person = types.model({
  *     name: types.string,
  *     gender: types.union(types.literal('male'), types.literal('female'))
  * })
- * 
+ *
  * @export
  * @alias types.literal
- * @template S 
+ * @template S
  * @param {S} value The value to use in the strict equal check
- * @returns {ISimpleType<S>} 
+ * @returns {ISimpleType<S>}
  */
 export function literal<S>(value: S): ISimpleType<S> {
     if (!isPrimitive(value)) fail(`Literal types can be built only on top of primitives`)
     return new Literal<S>(value)
 }
+
+/**
+ * The type of the value `null`
+ *
+ * @export
+ * @alias types.null
+ */
+export const nullType = literal(null)
+
+/**
+ * The type of the value `undefined`
+ *
+ * @export
+ * @alias types.undefined
+ */
+export const undefinedType = literal(undefined)